remove ja locale because it doesn't normalize caps, i.e. "Vocaloid" and "vocaloid" are different tags in ja locale
note: this never happened in rcv1 because in rcv1 the locale is "jp"
which is incorrect and is just treated as english lmao

not using a cjk locale is ok here because we have infix search, which
is very much a 'we have cjk at home' situation. works well enough,
but the main thing we lose is normalization (e.g. turning traditional
characters into simplified for search, or katakana to furigana etc)